requirements.txt
Flask>=2.3.2
Flask_HTTPAuth>=4.8.0
pandas>=2.0.2
prettytable>=3.8.0
requests>=2.31.0
selenium>=4.10.0
Werkzeug>=2.3.6

fastapi~=0.97.0
bcrypt~=4.0.1
uvicorn~=0.22.0
pydantic~=1.10.9
PyYAML~=6.0
redis~=4.5.5
matplotlib~=3.7.1
scikit-learn~=1.2.2
nltk~=3.8.1